feat: Extended Fluid class

+++ include/Particle.hpp
@@ -7,7 +7,7 @@ struct Particle {
 	glm::vec3 position;
 	glm::vec3 velocity;
 	glm::vec3 acceleration;
-	double mass;
+	float mass;
 };
